TST: travis: Expose system packages for virtualenv tests
In order to test the VenvEnvironment's upcoming system_site_packages
attribute on Travis, we need to expose Travis's system site packages.
Otherwise tests run in an isolated environment and test_venv.py's
attempt to create an environment with 'virtualenv --system-site-packages'
would have no effect.

Note that this only works if we are testing the same Python version as
the system Python on Travis (currently Python 3.4).