TST: resource.singularity: Use a more unique instance name
When we test Singularity instances, we set up the instances as usual,
using the same registry as any other instances created by that
user (through reproman or directly).  This means the state of the
running instances can affect the tests.  Ideally we could use an
isolated registry, but let's at least choose a more unique instance
name to make it unlikely that we collide with the user's other
instances.