ENH+TST: distribution.singularity: Pull by md5sum rather than commit
SingularityHub URLs support both commits and m5sums (though
Singularity v3 has a bug in pulling these URLs).  Switch from commit
hash to md5sum because using the md5sum makes it clearer to the reader
where the value is coming from and makes it easier to determine and
update since we can generate it from the image file alone.