TST: resource.singularity: Try to clean up instance on failure
As described in the previous commit, our tests use the global instance
registry.  If a check fails after the instance is created but before
it is deleted, we leave the test instance as a registered instance.

Let's try to clean up by doing the deletion even if any of the checks
between resource creation and deletion fail.