TST: resource.singularity: Drop extension check
Singularity v3 switched to Singularity Image Format as the default
image format, and by default 'singularity pull' will download images
to a file that ends in ".sif".  We could update this check to be
conditional on the version, but instead remove it because the next
commit will switch to using a local path for the image parameter
anyway.