NF: run: Support specification of batch parameters inline
As an alternative to --batch-spec, add an option for constructing the
batch parameters from combining lists of values.  Right now, we only
support taking the product of these values, but we could consider
adding a flag that specifying that each ith element should be
paired.  (See code comment for what would need to be adjusted.)

This syntax is pretty limited, but for more complicated things the
caller can use the --batch-spec option.