TST: travis: Use py3.5 for REPROMAN_TESTS_ASSUME_SSP build
As explained in c3ee9349c (TST: travis: Expose system packages for
virtualenv tests, 2019-06-05), the REPROMAN_TESTS_ASSUME_SSP build
needs to be matched up with the default Python version on Travis.

The default distribution for Travis has recently changed from Trusty
to Xenial, so the py3.4/REPROMAN_TESTS_ASSUME_SSP is failing [0].  Use
py3.5 for the REPROMAN_TESTS_ASSUME_SSP, which should come built in.
(And switch an existing py3.6 run to py3.4 so that we still have a run
for 3.4, which we claim to support.)

[0]: https://travis-ci.org/ReproNim/reproman/jobs/545343359#L194