ENH: resource: Provide more informative error about underscores
A common mistake when specifying the resource type is to use
underscores as the delimiter instead of hyphens (e.g.,
docker_container instead of docker-container).  This is an especially
forgivable mistake considering _we_ used the underscore form in the
output of backend-parameters and near-hit suggestions until a couple
of commits ago.

When the underscore variant is used, the caller sees an uninformative:

 Failed to find Docker_container in <module
 'reproman.resource.docker_container> [...]

Instead show

  'docker_container' not a known backend. Did you mean
  'docker-container'?

Another solution would be to accept either, but we'd need to make
wider changes to make that work properly.  Plus it's probably less
confusing to stick with one accepted form.