TST: Expose session fixture for resource manager
In some tests, we get a resource by calling ResourceManager.factory()
with a config dictionary.  That works fine because factory() is a
static method and there's no risk of interacting with the on-disk
inventory.

An upcoming commit will change factory() from a static method to a
regular one, so let's replace the above call with calls to a
session-scoped fixture that points to a temporary inventory path.