ENH: resource: Tolerate unknown resource parameters in inventory
It make sense to be strict when creating a resource, but when loading
something that has already made it into the inventory, let's be more
forgiving by issuing a warning and then dropping the unknown parameter
for instantiation.

The user is then responsible for cleaning up the value manually to get
rid of the warnings.  Instead of requiring manual clean-up, we could
modify .inventory and call save_inventory so that the key is dropped
from the on-disk inventory as well, but that is probably too
aggressive and opens up the possibility of destructively modifying
bits of the inventory due to bugs.

Closes #171.