ENH: orchestrators: Allow submit_command to be overridden
The immediate use for this is that I need to use a command other than
qsub for the PBS system I have access to.  More thought should be
given to making the submit command customizable, but for now just look
in the job spec for the (undocumented) "submit_command" key.  Note
this will working only from --job-spec, not --job-parameter, because
it needs to come in as a list.