RF: orchestartors: Adjust helper for cwd-enabled execute_command()
As of fd9bf948a (Added session handling of 'cwd' when executing a
command, 2019-03-21), we can use the `cwd` parameter.  Note that, if
for some reason we need to revert this change, shlex.quote() should be
used when constructing the prefix.

Keep _execute_in_wdir() around rather than calling execute_command()
directly because the helper is still useful for recasting the
CommandError's as OrchestratorError's.