BF: orchestrators: Try to put subdatasets on correct commit
We call `datalad {create-sibling,publish} --recursive` and, aside from
checking out the expected commit in the top-level dataset, behave as
though the hierarchy is in the desired state, but that's not
necessarily the case.  It isn't even always the case for a fresh setup
because publishing a subdataset with a branch other than master
checked out leaves the subdataset on an unborn master branch with an
empty working tree [*].

Go through each of the installed subdatasets and check out the
recorded commit.  Note that this is wasteful because it executes a
command on the resource even if the subdataset is already in the
desired state, but detecting whether the checkout is necessary would
only add to the problem as we'd have to execute an additional command.
But this workaround will hopefully be short-lived because we should be
able to drop this code once DataLad implements the ability to reset an
entire hierarchy to a desired state (issue 1424).

Closes #444.

[*] To add the confusion, Git doesn't mark this subdataset as
    modified (likely a bug in Git), which means our dirty check
    doesn't catch the unintended state.