BF: conda: Temporary pin version in miniconda URL
Our test suite is failing with the latest miniconda version (gh-443).
The initial issue looks to be a version conflict.  A reasonable
approach would be to just bump all versions in that test because
understanding or resolving this specific set of incompatibilities
between old software versions isn't worth the time or effort.

But that won't work because our distributions.conda isn't compatible
with the latest conda in at least two ways.  We use an environment
file that doesn't have an extension that conda now expects.  That's an
easy fix, and there's a patch in gh-443.  The bigger issue is that
environments seem to no longer have their own bin/conda unless conda
is explicitly installed in the environment.  (This mirrors the then
seemingly hard-to-trigger problem initially described in gh-173.)  As
a result, our method for identifying an environment's root prefix is
broken.  Based on a quick scan of envs/NAME/, there may no longer be
an easy way to map from an environment to its root prefix.

Given that there doesn't seem to be a quick fix and an always failing
test suite makes it hard to spot other, unknown failures, let's get
back to a passing state by pinning the miniconda version that we
download.