TST: orchestrators: Add check_orc_plain for SSH resource
We call this on a local and Docker resource.  Add a call for an SSH
resource because we have a reproman-test host on Travis as of
c4c710b7d (TST: travis: Set up SSH for localhost, 2019-08-08).

Note that for now we just skip the test if REPROMAN_TESTS_SSH is
unset, but we might want to be more specific because there are other
SSH-related tests that don't require having reproman-test set up.