TST: orchestrators: Add ssh fixture
We have a reproman-test host on Travis as of c4c710b7d (TST: travis:
Set up SSH for localhost, 2019-08-08).  We should switch most of our
orchestrator tests that use the local shell resource fixture over to
the ssh one because the local shell has a different code path for
prepare_remote() and fetch() because `datalad create-sibling` doesn't
currently support local paths.

Note that for now we just skip the test if REPROMAN_TESTS_SSH is
unset, but we might want to be more specific because there are other
SSH-related tests that don't require having reproman-test set up.