ENH: datalad-pair-run orchestrator: Reset to original commit
We used to do this, but f678ff746 (ENH: post-run template: Don't reset
repo state, 2018-11-14) removed it, claiming that we can't do that if
we want to fetch a detached run.  That is no longer true as of
7ad7218cb (ENH: orchestrators: Move HEAD in datalad-pair's fetch,
2019-03-04).

And the problem is that _not_ resetting puts the pair of repos under
the datalad-pair-run orchestrator into an ugly and hard to reconcile
state because the remote branch advances by the plain "datalad add ."
call while the local branch advances with a run commit.  The tests
didn't trigger an issue with this before because they used the unusual
localshell preparation, but we'll run into it as we switch over to
user the reproman-test ssh host.