BF: orchestrators: Fix parsing status from ls-tree output
Because the HEAD of the remote repo may change, we look for the status
in the jobid ref's tree if the corresponding metadata directory
doesn't exist.  This doesn't work as intended because we try to
convert the entire ls-tree line to an int rather than just the file
name (which is the subjob number).

Note that the adjusted test doesn't trigger the failure yet.  It needs
the next commit.