BF: orchestrators: Make dirty check work better with subdatasets
The dirty check in the remote repository asserts that there are no
uncommitted changes, including changes to recorded submodules.  This
works when the hierarchy is being created initially, but it can easily
fail if an existing submodule is updated.  For example, if a submodule
is switched to a branch other than master and that state is committed,
then `datalad publish` will result in a dirty state on the remote, but
that's a state we need to be able to work with.

Aside from waiting for the planned `datalad load` command that sets
the state of an entire hierarchy, there are a few solutions:

  * Assess the status of the entire hierarchy with `datalad
    status`. The main problem with this is that the command exists
    only in a developmental version of DataLad.

  * Use `git status --porcelain=v2` and don't complain about a dirty
    repo if the only output is for submodules that have a change in
    the checked out commit (i.e. the "sub" field is "SC..").  The
    problem with this is that it only works for one layer of
    submodules: if submodule A contains submodule B that has a
    different commit checked out than the one submodule A has
    recorded, the top-level repo shows submodule A as having tracked
    changes ("S.M.").

  * Do the dirty check, ignoring submodules, in each installed
    subdataset of the hierarchy.  The downside of this is that it adds
    a call for each dataset and will be slow when there are many
    repositories.

Go with the last one until for now.