TST: orchestrators: Drop module-scoped dataset fixture
818c71b20 (TST: orchestrators: Reduce datatset creation, 2019-03-04)
added a module-scoped fixture to reduce dataset creation.  As noted in
its commit message, the clean-up is too simplistic for many scenarios,
and it is quite easy to leak state.  Our cases are going to get more
complicated, so let's go back to creating the dataset in the
function-scoped fixture.  If performance ends up being an issue, we
can consider going to an approach that uses reckless clones.