ENH: orchestrators: Dump job spec in meta_directory
The local registry keeps track of the job spec after submission, but
removes it once the job is fetched into the local dataset.  However,
storing it with the other run metadata would be nice for
inspection (both for debugging and general information).  When the job
is fetched, the registry could inject it into the directory, but
tracking it in git would require modifying the history.  Instead let's
copy the spec to the remote machine *before* submission, as we do with
other metadata files.

Dumping the spec also opens up the possibility of rerunning a job from
a previous spec with `reproman run --js <spec>`.  At the moment,
feeding in a previous spec without filtering out some values won't
quite work; the next commits will address those issues.

Note the submission ID value in the spec dumped on the remote will
differ from the one in the local registry spec.  For the remote spec,
it will be None because we create the file on the remote before
submitting the command.  We could submit then dump the spec, but then
it'd race with the submitted script, just as writing the submission ID
to the remote idmap file currently does.  As with idmap creation, spec
creation losing the race isn't critical, but we might was well keep
the racy operation restricted to the idmap write.