ENH: run: Store resolved batch parameters under new key
If we overwrite "batch_parameters", the mapping rather than the list
of strings will be dumped as part of the spec.  If a caller tries to
execute another run by passing that dumped spec to --job-spec, the run
will abort saying that the batch_parameters value doesn't have the
expected "key=val1,val2,..." form.

Store the resolved value under a new key to avoid this.  Prefixing the
key with an underscore is in line with the changes that will come in
the next commit.