BF: orchestrators: Swap precedence of as_dict() sources
as_dict() is called to generate an object with job information to dump
to a yaml file for the local registry.  The two sources of values are
the keywords passed to the Template class---which includes every key
in the job spec---and a few object attributes (e.g., .resource.id).

The template keywords take precedence.  This currently doesn't really
matter because we're unlikely to encounter scenarios where the sets of
keys overlap, let alone cases where they disagree.  But upcoming
commits will make it easier to reuse a previous run's job spec that
was dumped with as_dict(), in which case all keys from the attribute
set will be in the template keywords set.  For example, the job spec
will come in with a resource_id value, and if that value and the
current .resource.id disagree, as_dict() would go with the previous
job spec's resource_id, despite the fact that the orchestrator class
actually uses .resource.id.

Make the attribute-based keys take precedence instead.