ENH: run: Use underscore for inaccessible job spec keys
The job spec contains keys that can be set by the user (through
--job-parameter and --job-spec) as well as inaccessible keys that are
set by us. The keys are now much more visible to the user now that we
dump the job spec as a metadata file, so it's more important that we
distinguish these two types of keys.  As a first step, let's prefix
keys that are inaccessible with an underscore.

We don't currently advertise that you can reuse the previous spec, but
where ever we end up doing that we should mention that what the
underscore prefix indicates.

The inaccessible keys correspond to keys that we generate during the
run, so we may want to just filter out keys that start with an
underscore right when the job specs come into run.py.