ENH: run: Don't touch command_str spec value
The user can specify command_str when supplying a job spec with
--job-spec.  Now that we dump the job spec as a metadata file, we
should avoid touching the value so that (1) the user isn't confused by
the seeing a value they didn't specify and (2) so that feeding a dump
spec to another run (via --job-spec) triggers the same processing as
the first.

Prefixing the key with an underscore is in line with the changes that
will come in the next commit.