TST: orchestrators: Add rudimentary test for reusing dumped spec
The test is limited because our orchestrators currently don't deal
well with existing files from the previous run in at least two ways:
the session.copy() method will fail when trying to copy input files
that already exist on a plain remote and, with a datalad remote,
specified outputs aren't unlocked.  Until the second issue is
resolved, we're only testing a case where we reset to the initial
HEAD.

Also, note that this test is working at the orchestrator class level,
so we don't yet have a check that interface/run.py can handle the
dumped job spec.