ENH: run: Version the dumped spec
As of 903b7aa8b (gh-464), we dump the spec into the metadata
directory.  One of the reasons for doing this is to make it possible
to use a dumped spec as the base for a spec fed back into the run().
Version the spec so that if we introduce incompatible changes to the
spec in the future, we can easily recognize old specs.