TST: run: Convert resource manager from module to function scope
The next commit introduces fetch actions, one of which is deleting a
resource.  To use this fixture there, we need to change the scope to
"function" to that the deletion doesn't affect other runs.  If this
ends up slowing down this module's tests too much, we can rework
things so that there is both a module- and function-scoped fixture.