MNT: Unpin docker-py
The docker module hasn't been released under the name "docker-py"
since 2016 [0] and has seen many releases under the name "docker" [1].
So we've effectively been pinning the release to version 1.10.6 even
though the current version is 4.1.0.

To update to latest version, we need to substitute APIClient for
Client (renamed in v2.3.0).  We also need to teach DockerSession.get()
about the new return value of APIClient.get_archive(): as of v3.0.0,
the content is returned as a generator of chunks rather than a byte
string.

Specify v3.0.0 as the new docker-py floor in setup.py

Closes #387.

[0]: https://pypi.org/project/docker-py
[1]: https://pypi.org/project/docker