ENH: docker: Pull :latest if no tag or digest is specified
If a tag or digest isn't specified when calling `reproman create`, we
download all tagged images.  This behavior will likely be confusing to
any callers that are familiar with the command-line interface to
docker pull, which defaults to :latest if a tag or digest is omitted.
And the behavior doesn't fit our goal: starting a new container from a
single image.

Teach the DockerContainer resource to append ":latest" when it
receives a bare image name.

Closes #483.