DOC: execute: Clarify sentence about git-tracked inputs
f6d36472f (DOC: a few minor tweaks and fixes to execute.rst,
2019-12-02) replaced "were" with "was", but that's incorrect if the
sentence is describing the hypothetical case where the file it tracked
by git-annex, not Git (which was the intention).  Rephrase the
sentence to fix the grammar and avoid confusion.

Re: https://github.com/ReproNim/reproman/pull/485#pullrequestreview-325325311