NF: run: Add slurm submitter
At this point, it should be on a par with the Condor submitter.  The
inline to-dos point to general things that need to be resolved across
the submitters (e.g., how to deal with discrepancy in batch script
parameters).

To test Slurm locally, our testing Slurm container can be put up with
the name "reproman-slurm-container" and a corresponding "slurm" host
can be added to the user's SSH configuration (see
tools/ci/setup-slurm-container.sh).

Note that we can't test the datalad-pair* orchestrators because the
testing container does not ship with DataLad.  Updating it to do so
will require more work because the base image of
giovtorres/docker-centos7-slurm has a very old Git (1.8.3.1) and
git-annex (5.20140221).

Closes #484.