RF: orchestrators: Move bulk of datalad checks to fixtures
This reorganization follows a similar pattern to what we already do
with check_orc_plain and will allow us to easily apply the same checks
to the Slurm submitter with the Slurm docker container as an SSH
resource rather than the localhost SSH resource that is used in
test_orc_datalad_run and test_orc_datalad_concurrent.

This change is mostly pure code movement, with no intended change in
what is tested.