BF: jobs: Don't dump job spec when logging orchestrators errors
When we started logging orchestrator errors in ab1c577c5 (RF: jobs:
Continue if action raises OrchestratorError, 2019-03-04), we put the
entire job spec into the log message.  That provides a flood of
information that is hard for the user to parse.  Replace the job spec
with the job ID, which was presumably the intended value in the first
place.