TST: run: Add local directory to context() fixture
The context() fixture calls run() from the top-level temporary
directory that is set to $HOME.  This is a bit unusual because the
local working directory for a run will almost certainly not be $HOME,
but it's not currently problematic for any of the tests.

The next commit, however, will delete the local working directory
immediately following the run() call.  As is, that would delete
$HOME/.reproman/run-root/ as part of the local working directory and
leave behind the process for `/bin/sh [...]/runscript 0`.

Move the local working directory to $HOME/local to avoid this issue.