TST: travis: Install released version of DataLad
Our build that pulls in DataLad is failing because we use the
deprecated `add` command, which has been removed entirely from
DataLad's master [*].  Let's get back to a passing state by installing the
latest release of DataLad.  We started installing DataLad's master in
0a2c30ff0 (TST: travis: Temporarily install developmental DataLad,
2018-11-20) to get access to functionality that hadn't made it into a
release yet.  That functionality is now in a release, so go back to
installing the released version.

[*] ReproMan needs an update for new DataLad functionality, which will
    include dropping our use of `add` and bumping our required DataLad
    version.