BF: orchestrators: Initialize submodules
In order to run something that involves submodules, we of course need
to populate them.  create-sibling won't do this for us as of DataLad's
78e00dcd23 (RF: do not run submodule update --init in the post-update
hook, 2019-04-10), so let's do a blanket 'update --init'.

This will probably need to be refined later, as getting all
subdatasets is unlikely to be desirable.  And we very likely should
try harder to stay on a branch.  This shouldn't functionally matter
with the upcoming 'update --follow=parent' change, but it risks
confusing users that inspect the remote directly.