ENH: orchestrators(datalad-pair): Do a more targeted update
datalad-pair's fetch() fetches the job ref from the resource and then
relies on 'datalad update' to bring down changes from the remote.  If
the merged in branch doesn't contain the job ref, then we temporarily
switch to the job ref, get the outputs, and then let the caller know
that the changes weren't brought in.

This approach is problematic with submodules.  One issue is that the
head switch mentioned above does not check out submodules, and in
general DataLad doesn't yet offer a way to set/reload a dataset
hierarchy to a particular state.

Another issue is that update() selects a merge target from the
submodule remote that doesn't necessarily contain the registered
commit.  See DataLad's 94efa8637a (NF: update: Add option to merge in
recorded submodule commit, 2020-02-17) for more information.  In what
will be the 0.13 release, update() learned a --follow=parentds mode
that makes update() take the registered commit as the merge target
when updating submodules.  Use it.

However, this still doesn't deal with bringing a non-mainline job ref
into the top-level dataset because update() unconditionally merges in
a branch from the remote.  Try harder to bring in the job ref by doing
an explicit 'git merge' of the ref before the update() call.