MNT: conda: Improve curl call for downloading miniconda script
The two important parts are (1) try to propagate failures rather than
failing silently and (2) follow redirects.  While here, disable
progress output as well.

Following the redirect isn't ideal because there won't be a failure to
alert us to a change in the URL.  However, not doing so means that, if
the URL changes, this functionality would be broken for a user until
we fix and the user upgrades, which is worse than us not realizing
there is a redirect.