RF: conda: Assume environments are in $root/envs
conda no longer installs conda in the bin/ of environments.  By
inspecting conda's source code, I haven't added luck tracing where
exactly this transition happened.  The method for running the conda
development isn't the same between two candidate endpoints for
bisection, so this is tricky to bisect in the repo.  (v4.3.31 is the
"installs conda within the environments" endpoint, based on the
version used in our tests).  And I'd rather not bisect manually with
downloaded miniconda's.

Make our tracer work with the new behavior by adding the assumption
that environments are under $root/envs/.  Then we can identify the
root by walking up, as we already do for virtual environments.  This
misses the (hopefully rare) case where environments are created with a
custom prefix, but at least we can work with environments that are
placed under the default location that is used when a name is
specified rather than a prefix.

Ref: https://github.com/ReproNim/reproman/issues/446#issuecomment-520064258