RF: conda: Avoid calling 'conda install' on root environment
For a reason that's not entirely clear, calling 'conda install' makes
the last install_packages() call in
test_conda_init_install_and_detect() fail with an error like

    Solving environment: ...working... failed
     err=
    ResolvePackageNotFound:
      - openssl==1.1.1f=h516909a_0
      - certifi==2020.4.5.1=py37hc8dfbb8_0
      - ca-certificates==2020.4.5.1=hecc5488_0

It may have something to do with not specifying channels, because the
test passes when the conda-forge channel is dropped from the root
environment.  But modifying the `install` call to pass the root
environment's channels with `--channel=` doesn't resolve the failure.

However, it's not clear why we need this 'conda install' call.  The
subsequent 'conda-env {update,create}' call should already handle this
for the root environment, so it seems that all that should matter is
that the root environment is created/updated before the other
environments.  So let's resolve the failure by removing the call to
'conda install' entirely and ensuring that the root environment gets
process first.

Ref: https://github.com/ReproNim/reproman/issues/446#issuecomment-612222453