TST: conda: Update some versions used in install test
We shouldn't do a blank update of every packages to its newer version,
because a part of the install_packages() functionality is that we
install the specified versions (even if that isn't explicitly tested
at the moment).  But let's at least bump the python, conda, and pip
versions.

Some combination of the conda and pip bump gets around a yaml.load()
error that is likely related to the deprecation of load(...,
Loader=None) in PyYAML 5.1.