TST: skip: Don't assume program is available in Runner-based checks
Most Runner.run() failures will result in a CommandError being raised.
However, if a command is given to run() in list form and a program
corresponding to the first argument isn't found, a FileNotFoundError
exception is raised instead.

This means that the no_condor() or no_slurm() checks _could_ crash (*)
if condor or docker, respectively, aren't installed on the system.
Update the Runner-based checks to catch a FileNotFoundError error.

(*) In practice, the failure in no_condor() isn't reachable because it
    only runs condor_status when external_versions reports that condor
    is available, and external_versions catches a blanket exception
    around its call to version functions.  And the no_slurm() failure,
    while accessible, is hidden a bit because it requires enabling
    REPROMAN_TESTS_SSH.