ENH: cmd: Log at debug on FileNotFoundError if expecting failure
If a command is given to run() in list form and a program
corresponding to the first argument isn't found, a FileNotFoundError
exception is raised.  This means that the failure is logged as an
error in many spots where we call run() and a failure isn't unexpected
(e.g., external_versions).

Instead log this case at the debug level if expect_fail=True.

Note that this code was originally from DataLad.  The runner there has
diverged quite a bit.  At this point, bringing in updates via 3rd is
likely to be painful (and in general hard to maintain), but if someone
attempts it, taking their side for this bit of code would be fine, as
the error is now unconditionally logged at 11.