ENH: external_versions: Call run expecting failure/stderr
As of a couple of commits back, expect_fail=True will result in a
FileNotFoundError exception being logged at the debug rather than the
error level.  Update the run calls to use expect_fail=True, as well as
expect_stderr=True while we're at it, because neither a failure to
find these programs nor stderr emitted while running them should be
presented as an error to the user.