ENH: orchestrators: Provide access to datalad-publish results
To help debug failures, make all results available at the DEBUG level,
and display status="error" results directly in the OrchestratorError
message.

At this point, we're not doing any analysis of the error message, so
remove the datalad-update suggestion.  As demonstrated by gh-441, this
may fail for a variety of reasons, and in some cases the
datalad-update snippet will just be confusing.  Eventually it might be
good to do more analysis on an error to give targeted suggestions.
And at the very least it'd probably be good to format the message file
if it's coming in as a (<format string>, <args>) tuple.

Closes #441.
Re: #511