BF: runscript - only consider statuses for success/failure
When running the usecase example gh-438, the working tree is coming up
as unexpectedly dirty, including the change of status files from empty
to "succeed", apparently indicating that the content of the file is
not flushed at the time of the 'datalad add/save' call.  To prevent
saving before the content is flushed, don't count a status file unless
it has the content.

Modified-by: Kyle Meyer <kyle@kyleam.com>
  Add commit message body, and dropped `find` call, as suggested by
  @yarikoptic.