ENH: run: Support using local directory as working directory
It's possible to run jobs on a local resource, but the local resource
is treated as any other resource.  A separate working directory is
prepared, and the results are fetched from there into the initial
"local" directory.

The above makes sense in the context of a `reproman run` that is
focused on remote resources, with local variants being useful for
testing.  However, if the goal is to actually promote `reproman run`
as something useful on local resources (gh-460), then the above
two-directory setup is unnecessary and inefficient.

Create a "datalad-no-remote" orchestrator that sets the working
directory to the local directory, pruning the unnecessary elements of
prepare_remote() and fetch().

Closes #460.