RF: orchestrators: Move repeated datalad result handling to helper
There are a few spots that do roughly the same inspection of datalad's
result records for errors, with the notable difference that some
should check for "impossible" but do not.  Move the logic to a common
helper.

Note that this change makes update() log failures, going against the
now-removed comment saying that datalad will handle it.  Failures
should be logged here (1) for consistency with what we do elsewhere
and (2) because datalad configuration can be used to hide datalad's
logging, so we shouldn't count on it.