ENH: orchestrators: Check URL before giving "sibling exists" error
If the target directory on the remote doesn't exist,
DataladOrchestrator.prepare_remote() checks whether a sibling of the
same name exists and aborts if it does.

As touched on in gh-440, this is stricter than necessary.  If the
sibling exists but points to the expected target directory on the
resource, we can proceed.  We just need to remove the remote first so
that the subsequent publish() call doesn't fail trying to fetch from a
non-existent remote.

Note that this doesn't fully address the request in gh-440, which asks
for reproman to offer to reconfigure the full entire hierarchy.  And
it only checks the top-level target.  Taken along with the improved
subdataset handling in 56b6de6c6 (BF: orchestrators: Call
create-sibling unconditionally, 2020-05-25), the change might be good
enough.  If not, a more extensive (and expensive) check of subdatasets
will be needed.