BF: orchestrators: Don't crash if run ref doesn't exist
We check for the status file within the tree of refs/reproman/$jobid
when the status file isn't found in the working tree, but it may not
exist at all if the runscript failed for other reasons (e.g., the
wrong parallel is installed, as in gh-522).

Catch the exception and let follow() give a more informative error, as
intended.