ENH: ls: Add --resref-type
a91fd0ffa (ENH: ls: Support limiting the output to specified
resources, 2019-02-26) left out the ability to disambiguate the
specified resources with --resref-type.  Providing that option is a
little bit odd because multiple resources can be provided and it would
apply to all of them.  On the other hand, it's hard to imagine what
else the behavior would be, and it's useful for a programatic check of
whether the resource exists (gh-521).