ENH: ls: Exit with non-zero code when given an unknown resource
Update Ls.__call__() to raise a ResourceNotFoundError if any call to
get_resource() raises a ResourceNotFoundError.  This makes it easier
for a command-line caller to check if a resource needs to be created
with `reproman ls --resref-type=name NAME`.

Note that there are other exceptions that we might want to map to
non-zero exit code (e.g., ambiguous resref or even a connection
error), in which case we'll need to move to using distinct exit codes
so that the above caller can still detect the "unknown resource" case.

Closes #521.