BF: condor.template: Revert should_transfer_files=Yes
ad85f1399 (ENH: Added options to HTCondor config template to handle
moving files across the cluster to/from the master node, 2020-02-27)
instructed condor to transfer files from and back to the submitting
machine.  This configuration results in two issues:

  * test_orc_datalad_run[sub:condor-orc:pair] fails on my local
    machine.  It completes its first job and then fails when preparing
    the second because the remote repository is unexpectedly dirty.
    The untracked files are the previous stdout and stderr that condor
    transfers back at exit, at which point the runscript has already
    switched away from the HEAD checked out to execute the job.

  * test_orc_datalad_concurrent[sub:condor-orc:pair-run] stalls for an
    unknown reason.

    https://travis-ci.org/github/ReproNim/reproman/jobs/714911219

As far as I understand, adding this configuration was a false start
for an issue that was then solved by 3d93c7e0a (ENH: Added NFS to AWS
HTCondor cluster management, 2020-04-06), so let's remove it.