TST: travis: Use bionic for --system-site-packages job
The --system-site-packages Travis job has started to fail in the 'make
-C docs html doctest' phase [0]

    Exception occurred:
      File "/usr/lib/python3.5/typing.py", line 1134, in _get_defaults
        code = func.__code__
    AttributeError: type object 'GitRunner' has no attribute '__code__'

The Python version used in that job is 3.5.2.  As of Python 3.5.3
(specifically CPython's 991d14fee1), that attribute error is caught
within typing._get_defaults().  That explains why this isn't seen in
the other builds (even those that use 3.5 use a later version), though
it doesn't explain why this started to fail only recently.  That might
have to do with the Sphinx; the appearance of the failure aligns with
the Sphinx 3.2.0 release.

Update the --system-site-packages to `dist: bionic` to get a Python
version with the fix.

[0] https://travis-ci.org/github/ReproNim/reproman/jobs/717707039#L1978

Fixes #542.